requirements.txt
tensorflow>=2.10.0,<2.16.0
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0,<1.25.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
